Separating the asynchronous configuration into its own class
